Moved welcome component into shared component
