+++ src/components/welcome/welcome.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../../variables";
+@import "../../variables.scss";
 
 .intro {
   margin: 0.5rem 0 1rem;
