--- src/pages/panel/components/welcomeCard/welcomeCard.module.scss
@@ -1,38 +1,0 @@
-@import "../../../../variables.scss";
-
-.background {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  position: fixed;
-  z-index: 10;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-
-  background: rgba(0, 0, 0, 0.5);
-}
-
-.close {
-  position: absolute;
-  top: 50%;
-  right: 0.25rem;
-  transform: translateY(-50%);
-  width: 2rem;
-
-  appearance: none;
-  display: block;
-  font-size: 1.5rem;
-  line-height: 1.1;
-  color: $primary-text;
-  cursor: pointer;
-  text-align: center;
-  transition: $transition color;
-
-  &:hover,
-  &:focus {
-    color: $outline-color;
-  }
-}
