+++ src/pages/panel/components/welcomeCardOverlay/welcomeCardOverlay.module.scss
@@ -1,0 +1,20 @@
+@import "../../../../variables.scss";
+
+.background {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  position: fixed;
+  z-index: 10;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+
+  background: rgba(0, 0, 0, 0.5);
+}
+
+.overlay {
+  position: relative;
+}
